Update functional.normalize
Array normalisation seems to be a surprisingly slow process. Using OpenCV instead of NumPy operations it's 30% faster.

Note that the proposed `normalize()` function does not return the same exact array. but the mean absolute difference is of the order of 5e-8

~~~~~~~~~~
Timings for the operation on an image of shape (1024, 1224, 3):
- OLD: 37.6 ms ± 767 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
- NEW: 22.4 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)