Fixing a Corner Case Bug in Compose when transforms=[] (#1782)
* There is a corner-case bug in `composition.py`: 
When `transforms` is `[]`, `_set_keys()` will NOT be adding `additional_targets` into `self._available_keys` (as it is just going through all transforms and processors, which is nothing done when there is no transform or processor), and causes runtime errors if passing those in the usage. 
Adding the L151 will fix this corner case.

* Update test_core.py

Add `test_compose_additional_targets_in_available_keys()` to test the corner case of the bug's occurrence.